Added Policies in Domains + new helper functions - Policies ensure authorization for specific actions - Users unauthorized will receive 401 error - Extended tests with differentiation in users and admins - Global helpers.php file - Language fixes - Removed views directory

+++ src/User/Application/Repositories/Eloquent/UserRepository.php
@@ -2,12 +2,12 @@
 
 namespace Src\User\Application\Repositories\Eloquent;
 
-use Illuminate\Database\Eloquent\ModelNotFoundException;
 use Illuminate\Support\Collection;
 use Src\User\Application\DTO\UserData;
 use Src\User\Application\Exceptions\EmailAlreadyUsedException;
 use Src\User\Domain\Model\User;
 use Src\User\Domain\Model\ValueObjects\Password;
+use Src\User\Domain\Policies\UserPolicy;
 use Src\User\Domain\Repositories\AvatarRepositoryInterface;
 use Src\User\Domain\Repositories\UserRepositoryInterface;
 use Src\User\Infrastructure\EloquentModels\UserEloquentModel;
@@ -15,14 +15,17 @@ use Src\User\Infrastructure\EloquentModels\UserEloquentModel;
 class UserRepository implements UserRepositoryInterface
 {
     private AvatarRepositoryInterface $avatarRepository;
+    private UserPolicy $userPolicy;
 
     public function __construct(AvatarRepositoryInterface $avatarRepository)
     {
         $this->avatarRepository = $avatarRepository;
+        $this->userPolicy = new UserPolicy();
     }
 
     public function findAll(): Collection
     {
+        authorize('findAll', $this->userPolicy);
         $users = collect();
         foreach (UserEloquentModel::all() as $userEloquent) {
             $users->push(UserData::fromEloquent($userEloquent, $this->avatarRepository));
@@ -32,18 +35,21 @@ class UserRepository implements UserRepositoryInterface
 
     public function findById(string $userId): User
     {
+        authorize('findById', $this->userPolicy);
         $userEloquent = UserEloquentModel::findOrFail($userId);
         return UserData::fromEloquent($userEloquent, $this->avatarRepository);
     }
 
     public function findByEmail(string $email): User
     {
+        authorize('findByEmail', $this->userPolicy);
         $userEloquent = UserEloquentModel::where('email', $email)->firstOrFail();
         return UserData::fromEloquent($userEloquent, $this->avatarRepository);
     }
 
     public function store(User $user, Password $password): User
     {
+        authorize('store', $this->userPolicy);
         if (UserEloquentModel::where('email', $user->email)->exists()) {
             throw new EmailAlreadyUsedException();
         }
@@ -63,6 +69,7 @@ class UserRepository implements UserRepositoryInterface
 
     public function update(User $user, Password $password, bool $updateAvatar): User
     {
+        authorize('update', $this->userPolicy, ['user' => $user]);
         $avatar = $user->avatar;
         if ($avatar->isBinaryFile()) {
             $filename = $this->avatarRepository->storeAvatarFile($avatar, $user->name);
@@ -82,6 +89,7 @@ class UserRepository implements UserRepositoryInterface
 
     public function delete(int $user_id): void
     {
+        authorize('delete', $this->userPolicy);
         $userEloquent = UserEloquentModel::findOrFail($user_id);
         $userEloquent->delete();
     }
